+++ v3/newrelic/app_run.go
@@ -144,10 +144,10 @@ func newAppRun(config config, reply *internal.ConnectReply) *appRun {
 
 	run.harvestConfig = harvestConfig{
 		ReportPeriods:   run.ReportPeriods(),
-		MaxTxnEvents:    run.MaxTxnEvents(),
-		MaxCustomEvents: run.MaxCustomEvents(),
-		MaxErrorEvents:  run.MaxErrorEvents(),
-		MaxSpanEvents:   run.MaxSpanEvents(),
+		MaxTxnEvents:    run.limit(run.Config.TransactionEvents.MaxSamplesStored, run.ptrTxnEvents),
+		MaxCustomEvents: run.limit(run.Config.CustomInsightsEvents.MaxSamplesStored, run.ptrCustomEvents),
+		MaxErrorEvents:  run.limit(run.Config.ErrorCollector.MaxSamplesStored, run.ptrErrorEvents),
+		MaxSpanEvents:   run.limit(run.Config.SpanEvents.MaxSamplesStored, run.ptrSpanEvents),
 		LoggingConfig:   run.LoggingConfig(),
 	}
 
@@ -221,23 +221,12 @@ func (run *appRun) ptrLogEvents() *uint    { return run.Reply.EventData.Limits.L
 func (run *appRun) ptrErrorEvents() *uint  { return run.Reply.EventData.Limits.ErrorEvents }
 func (run *appRun) ptrSpanEvents() *uint   { return run.Reply.SpanEventHarvestConfig.HarvestLimit }
 
-func (run *appRun) MaxTxnEvents() int { return run.limit(run.Config.maxTxnEvents(), run.ptrTxnEvents) }
-func (run *appRun) MaxCustomEvents() int {
-	return run.limit(internal.MaxCustomEvents, run.ptrCustomEvents)
-}
-func (run *appRun) MaxLogEvents() int {
-	return run.limit(internal.MaxLogEvents, run.ptrLogEvents)
-}
-func (run *appRun) MaxErrorEvents() int {
-	return run.limit(internal.MaxErrorEvents, run.ptrErrorEvents)
-}
-
 func (run *appRun) LoggingConfig() (config loggingConfig) {
 	logging := run.Config.ApplicationLogging
 
 	config.loggingEnabled = logging.Enabled
 	config.collectEvents = logging.Enabled && logging.Forwarding.Enabled && !run.Config.HighSecurity
-	config.maxLogEvents = run.MaxLogEvents()
+	config.maxLogEvents = run.limit(logging.Forwarding.MaxSamplesStored, run.ptrLogEvents)
 	config.collectMetrics = logging.Enabled && logging.Metrics.Enabled
 	config.localEnrichment = logging.Enabled && logging.LocalDecorating.Enabled
 	if run.Config.Labels != nil && logging.Forwarding.Enabled && logging.Forwarding.Labels.Enabled {
@@ -250,18 +239,11 @@ func (run *appRun) LoggingConfig() (config loggingConfig) {
 	return config
 }
 
-// MaxSpanEvents returns the reservoir limit for collected span events,
-// which will be the default or the user's configured size (if any), but
-// may be capped to the maximum allowed by the collector.
-func (run *appRun) MaxSpanEvents() int {
-	return run.limit(internal.MaxSpanEvents, run.ptrSpanEvents)
-}
-
-func (run *appRun) limit(dflt int, field func() *uint) int {
+func (run *appRun) limit(configMaxSamplesStored int, field func() *uint) int {
 	if field() != nil {
 		return int(*field())
 	}
-	return dflt
+	return configMaxSamplesStored
 }
 
 func (run *appRun) ReportPeriods() map[harvestTypes]time.Duration {