+++ v3/newrelic/config.go
@@ -113,6 +113,8 @@ type Config struct {
 		ExpectStatusCodes []int
 		// Attributes controls the attributes included with errors.
 		Attributes AttributeDestinationConfig
+		// Reservoir limit for error events. Defaults to 100
+		MaxSamplesStored int
 		// RecordPanics controls whether or not a deferred
 		// Transaction.End will attempt to recover panics, record them
 		// as errors, and then re-panic them.  By default, this is
@@ -304,6 +306,9 @@ type Config struct {
 		Enabled bool
 		// Attributes controls the attributes included on Spans.
 		Attributes AttributeDestinationConfig
+		// MaxSamplesStored allows you to limit the number of Span
+		// Events stored/reported in a given 60-second period
+		MaxSamplesStored int
 	}
 
 	// InfiniteTracing controls behavior related to Infinite Tracing tail based
@@ -660,6 +665,7 @@ func defaultConfig() Config {
 	c.TransactionEvents.Enabled = true
 	c.TransactionEvents.Attributes.Enabled = true
 	c.TransactionEvents.MaxSamplesStored = internal.MaxTxnEvents
+
 	c.HighSecurity = false
 	c.ErrorCollector.Enabled = true
 	c.ErrorCollector.CaptureEvents = true
@@ -670,6 +676,7 @@ func defaultConfig() Config {
 		http.StatusNotFound, // 404
 	}
 	c.ErrorCollector.Attributes.Enabled = true
+	c.ErrorCollector.MaxSamplesStored = internal.MaxErrorEvents
 	c.Utilization.DetectAWS = true
 	c.Utilization.DetectAzure = true
 	c.Utilization.DetectPCF = true
@@ -707,6 +714,7 @@ func defaultConfig() Config {
 	c.DistributedTracer.Sampler.RemoteParentNotSampled = Default.String()
 	c.SpanEvents.Enabled = true
 	c.SpanEvents.Attributes.Enabled = true
+	c.SpanEvents.MaxSamplesStored = internal.MaxSpanEvents
 
 	c.DatastoreTracer.InstanceReporting.Enabled = true
 	c.DatastoreTracer.DatabaseNameReporting.Enabled = true
@@ -802,30 +810,45 @@ func (c Config) validateTraceObserverConfig() (*observerURL, error) {
 	}, nil
 }
 
-// maxTxnEvents returns the configured maximum number of Transaction Events if it has been configured
-// and is less than the default maximum; otherwise it returns the default max.
-func (c Config) maxTxnEvents() int {
-	configured := c.TransactionEvents.MaxSamplesStored
+// maxTxnEvents returns the configured maximum number of Transaction Events if it
+// is less than the default maximum; otherwise it returns the default max.
+func maxTxnEvents(configured int) int {
 	if configured < 0 || configured > internal.MaxTxnEvents {
 		return internal.MaxTxnEvents
 	}
 	return configured
 }
 
-// maxCustomEvents returns the configured maximum number of Custom Events if it has been configured
-// and is less than the default maximum; otherwise it returns the default max.
-func (c Config) maxCustomEvents() int {
-	configured := c.CustomInsightsEvents.MaxSamplesStored
+// maxSpanEvents returns the configured maximum number of Span Events if it
+// is less than the default maximum; otherwise it returns the default max.
+func maxSpanEvents(configured int) int {
+	if configured < 0 || configured > internal.MaxSpanEvents {
+		return internal.MaxSpanEvents
+	}
+	return configured
+}
+
+// maxCustomEvents returns the configured maximum number of Custom Events if it
+// is less than the default maximum; otherwise it returns the default max.
+func maxCustomEvents(configured int) int {
 	if configured < 0 || configured > internal.MaxCustomEvents {
 		return internal.MaxCustomEvents
 	}
 	return configured
 }
 
-// maxLogEvents returns the configured maximum number of Log Events if it has been configured
-// and is less than the default maximum; otherwise it returns the default max.
-func (c Config) maxLogEvents() int {
-	configured := c.ApplicationLogging.Forwarding.MaxSamplesStored
+// maxErrorEvents returns the configured maximum number of Error Events if it
+// is less than the default maximum; otherwise it returns the default max.
+func maxErrorEvents(configured int) int {
+	if configured < 0 || configured > internal.MaxErrorEvents {
+		return internal.MaxErrorEvents
+	}
+	return configured
+}
+
+// maxLogEvents returns the configured maximum number of Log Events if it
+// is less than the default maximum; otherwise it returns the default max.
+func maxLogEvents(configured int) int {
 	if configured < 0 || configured > internal.MaxLogEvents {
 		return internal.MaxLogEvents
 	}
@@ -1014,7 +1037,7 @@ func configConnectJSONInternal(c Config, pid int, util *utilization.Data, e envi
 		Util:             util,
 		SecurityPolicies: securityPolicies,
 		Metadata:         metadata,
-		EventData:        internal.DefaultEventHarvestConfigWithDT(c.TransactionEvents.MaxSamplesStored, c.ApplicationLogging.Forwarding.MaxSamplesStored, c.CustomInsightsEvents.MaxSamplesStored, c.DistributedTracer.ReservoirLimit, c.DistributedTracer.Enabled),
+		EventData:        internal.DefaultEventHarvestConfigWithDT(c.TransactionEvents.MaxSamplesStored, c.ApplicationLogging.Forwarding.MaxSamplesStored, c.CustomInsightsEvents.MaxSamplesStored, c.SpanEvents.MaxSamplesStored, c.DistributedTracer.Enabled),
 	}})
 }